refactor: 💡 refactoring contracts && invoices of sales screens
